Add StarDOP support, via SU660
Yeah, this is a kludgy merge, but nvidia doesn't even bother updating
their kernel repositories anymore. FML...

+++ include/linux/mmc/sdhci-pltfm.h
@@ -27,7 +27,7 @@ struct sdhci_host;
  */
 struct sdhci_pltfm_data {
 	struct sdhci_ops *ops;
-	unsigned int quirks;
+	u64 quirks;
 	int (*init)(struct sdhci_host *host, struct sdhci_pltfm_data *pdata);
 	void (*exit)(struct sdhci_host *host);
 };
